+++ CHANGELOG.md
@@ -70,6 +70,7 @@ it in future.
 
 ### Fixed
 
+* Remove SIMPATH dependency, replaced with EVTGENDATA for EvtGen data files (#648)
 * fix(digi): Make TTree branch split level configurable in BaseDetector, set splitLevel=1 for MTC
 * chore: Fix file endings
 * chore: trim trailing whitespace